Separate SGX error handling into a new file
This changes is done to avoid SGX dependecy by any module which just needs
to use general error handling functionalities provided by error.h
