+++ tc/sgx/trusted_worker_manager/enclave/base_enclave.cpp
@@ -28,6 +28,7 @@
 #include "auto_handle_sgx.h"
 
 #include "error.h"
+#include "avalon_sgx_error.h"
 #include "tcf_error.h"
 #include "zero.h"
