+++ tc/sgx/trusted_worker_manager/enclave/enclave_data.cpp
@@ -22,6 +22,7 @@
 
 #include "crypto.h"
 #include "error.h"
+#include "avalon_sgx_error.h"
 #include "tcf_error.h"
 #include "zero.h"
